+++ changelog.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2025-04-13
+
+### Added
+- Integrated support for new LLM providers:
+  - `MoonshotAI (Kimi)`
+  - `Cloudflare` (available in the private mode only: temporary unavailable in the public mode)
+
+### Changed
+- Project dependencies updated.
+
 ## [1.0.0] - 2025-04-01
 
 ### Added